refactor: 修改了代码的格式，删除了多余的文件
